refactor: 그래픽 이미지 변경
