+++ src/pages/host/components/index.ts
@@ -1,5 +1,6 @@
 import HostMyClassCard from './HostMyClassCard/HostMyClassCard';
 import ApplicantListModal from './ApplicantListModal/ApplicantListModal';
 import ClassManageEmptyView from './ClassManageEmptyView/ClassManageEmptyView';
+import HostMyClassEmptyView from './HostMyClassEmptyView/HostMyClassEmptyView';
 
-export { HostMyClassCard, ApplicantListModal, ClassManageEmptyView };
+export { HostMyClassCard, ApplicantListModal, ClassManageEmptyView, HostMyClassEmptyView };
